backend/requirements.txt
Flask==2.2.2
openai==1.6.1
azure-storage-blob
azure-identity
Werkzeug==2.3.7
redis==3.5.3
